+++ src/FibonacciWebApi/Services/IFibonacciService.cs
@@ -2,12 +2,12 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-
+using System.Numerics;
 namespace FibonacciWebApi.Services
 {
      public interface IFibonacciService
     {
-         long CalcularFibo(int num);
+         BigInteger CalcularFibo(int num);
 
     }
 }
